+++ core/smd-api/src/main/java/app/dodb/smd/api/event/channel/EventChannel.java
@@ -1,9 +1,4 @@
 package app.dodb.smd.api.event.channel;
 
-import app.dodb.smd.api.event.Event;
-import app.dodb.smd.api.event.EventMessage;
-
-public interface EventChannel {
-
-    <E extends Event> void send(EventMessage<E> eventMessage);
+public interface EventChannel extends EventSink, EventSource {
 }