+++ core/smd-api/src/main/java/app/dodb/smd/api/event/channel/EventSink.java
@@ -1,0 +1,9 @@
+package app.dodb.smd.api.event.channel;
+
+import app.dodb.smd.api.event.Event;
+import app.dodb.smd.api.event.EventMessage;
+
+public interface EventSink {
+
+    <E extends Event> void send(EventMessage<E> eventMessage);
+}