+++ core/smd-api/src/main/java/app/dodb/smd/api/event/channel/EventSource.java
@@ -1,6 +1,6 @@
 package app.dodb.smd.api.event.channel;
 
-public interface SubscribableEventChannel extends EventChannel {
+public interface EventSource {
 
     void subscribe(EventChannelListener listener);
 }