+++ examples/ticket-service/src/integrationTest/java/app/dodb/smd/ticket/drivenadapter/ticketactivity/TicketActivityIntegrationTest.java
@@ -4,7 +4,7 @@ import app.dodb.smd.api.event.EventMessage;
 import app.dodb.smd.api.event.EventPublisher;
 import app.dodb.smd.api.event.bus.ProcessingGroupsConfigurer;
 import app.dodb.smd.api.metadata.Metadata;
-import app.dodb.smd.eventstore.channel.EventStoreChannel;
+import app.dodb.smd.eventstore.channel.EventStore;
 import app.dodb.smd.ticket.drivenadapter.IntegrationTest;
 import app.dodb.smd.ticket.drivingport.event.TicketAssignedEvent;
 import app.dodb.smd.ticket.drivingport.event.TicketOpenedEvent;
@@ -35,9 +35,9 @@ class TicketActivityIntegrationTest {
     static class TestConfig {
 
         @Bean
-        ProcessingGroupsConfigurer ticketProcessingGroups(EventStoreChannel eventStoreChannel) {
+        ProcessingGroupsConfigurer ticketProcessingGroups(EventStore eventStore) {
             return spec -> spec
-                .processingGroup("ticket-activity").channel(eventStoreChannel)
+                .processingGroup("ticket-activity").source(eventStore)
                 .anyProcessingGroup().disabled();
         }
     }