+++ framework/smd-spring-boot-starter/src/integrationTest/java/app/dodb/smd/spring/eventstore/processing/EventStoreProcessingTestConfiguration.java
@@ -2,7 +2,7 @@ package app.dodb.smd.spring.eventstore.processing;
 
 import app.dodb.smd.api.event.ProcessingGroup;
 import app.dodb.smd.api.event.bus.ProcessingGroupsConfigurer;
-import app.dodb.smd.eventstore.channel.EventStoreChannel;
+import app.dodb.smd.eventstore.channel.EventStore;
 import app.dodb.smd.spring.EnableSMD;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.Bean;
@@ -13,7 +13,7 @@ import javax.sql.DataSource;
 @Configuration
 @EnableAutoConfiguration
 @EnableSMD
-public class EventStoreChannelProcessingTestConfiguration {
+public class EventStoreProcessingTestConfiguration {
 
     @Bean
     public FailableTestEventHandler failableTestEventHandler() {
@@ -26,8 +26,8 @@ public class EventStoreChannelProcessingTestConfiguration {
     }
 
     @Bean
-    public ProcessingGroupsConfigurer processingGroupsConfigurer(EventStoreChannel eventStoreChannel) {
+    public ProcessingGroupsConfigurer processingGroupsConfigurer(EventStore eventStore) {
         return spec -> spec
-            .processingGroup(ProcessingGroup.DEFAULT).channel(eventStoreChannel);
+            .processingGroup(ProcessingGroup.DEFAULT).source(eventStore);
     }
 }