+++ framework/smd-spring-boot-starter/src/main/java/app/dodb/smd/spring/CommandBusSpecCustomizer.java
@@ -1,0 +1,9 @@
+package app.dodb.smd.spring;
+
+import app.dodb.smd.api.command.bus.CommandBusSpec;
+
+@FunctionalInterface
+public interface CommandBusSpecCustomizer {
+
+    void customize(CommandBusSpec spec);
+}