+++ framework/smd-spring-boot-starter/src/main/java/app/dodb/smd/spring/EventBusSpecCustomizer.java
@@ -1,0 +1,9 @@
+package app.dodb.smd.spring;
+
+import app.dodb.smd.api.event.bus.EventBusSpec;
+
+@FunctionalInterface
+public interface EventBusSpecCustomizer {
+
+    void customize(EventBusSpec spec);
+}