+++ framework/smd-spring-boot-starter/src/main/java/app/dodb/smd/spring/QueryBusSpecCustomizer.java
@@ -1,0 +1,9 @@
+package app.dodb.smd.spring;
+
+import app.dodb.smd.api.query.bus.QueryBusSpec;
+
+@FunctionalInterface
+public interface QueryBusSpecCustomizer {
+
+    void customize(QueryBusSpec spec);
+}